Change DrsObject.size to long, and associated methods

+++ src/main/java/org/ga4gh/drs/model/DrsObject.java
@@ -32,7 +32,7 @@ public class DrsObject {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'")
     private LocalDateTime createdTime;
 
-    private int size;
+    private long size;
 
     // Optional
     private List<AccessMethod> accessMethods;
@@ -58,7 +58,7 @@ public class DrsObject {
         
     }
 
-    public DrsObject(String id, URI selfURI, List<Checksum> checksums, LocalDateTime createdTime, int size) {
+    public DrsObject(String id, URI selfURI, List<Checksum> checksums, LocalDateTime createdTime, long size) {
         this.setId(id);
         this.setSelfURI(selfURI);
         this.setChecksums(checksums);
@@ -98,11 +98,11 @@ public class DrsObject {
         this.createdTime = createdTime;
     }
 
-    public int getSize() {
+    public long getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(long size) {
         this.size = size;
     }
